+++ src/pages/index.tsx
@@ -3,6 +3,7 @@ import * as React from 'react'
 import Layout from '@/components/layout/Layout'
 import CV from '@/components/MainPage/CV'
 import Introductory from '@/components/MainPage/Introductory'
+import ProjectList from '@/components/MainPage/ProjectList'
 import Seo from '@/components/Seo'
 
 export default function Index(): JSX.Element {
@@ -12,6 +13,7 @@ export default function Index(): JSX.Element {
 			<div className='container py-2 md:px-36 dark:text-white'>
 				<Introductory />
 				<CV />
+				<ProjectList />
 			</div>
 		</Layout>
 	)