+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - **dark-mode:** dark mode now persistent via local storage
 - **main-page:** added profile picture
 - **main-page:** added introductory text
+- **project-diplay-card:** added card
 
 ### Fix
 - **changelog:** fixed changelog format